requirements.txt
djangorestframework
django-cors-headers
python-dotenv
djangorestframework-simplejwt
django-filter
Pillow
whitenoise
psycopg2-binary